wip simplify expression resolver
