+++ sql/column_spec.go
@@ -1,0 +1,24 @@
+// Copyright 2019 The SQLFlow Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sql
+
+// columnSpec defines how to generate codes to parse column data to tensor/sparsetensor
+type columnSpec struct {
+	ColumnName string
+	IsSparse   bool
+	Shape      []int
+	DType      string
+	Delimiter  string
+	FeatureMap featureMap
+}
